user_authn: Move password-related tests to test_password.py
